backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.4
python-dotenv==1.0.1
supabase==2.8.1
psycopg2-binary
sqlalchemy
python-multipart
